(misc) fix windows build after earlier refactor
