+++ cmd/server_run.go
@@ -1,11 +1,8 @@
-// +build !windows
-
 package cmd
 
 import (
 	"fmt"
 	"os"
-	"sync"
 
 	"github.com/choria-io/go-choria/choria"
 	"github.com/choria-io/go-choria/config"
@@ -65,26 +62,6 @@ func (r *serverRunCommand) Configure() error {
 	}
 
 	return nil
-}
-
-func (r *serverRunCommand) Run(wg *sync.WaitGroup) (err error) {
-	defer wg.Done()
-
-	if len(c.BuildInfo().AgentProviders()) == 0 {
-		return fmt.Errorf("invalid Choria Server build, no agent providers present")
-	}
-
-	instance, err := r.prepareInstance()
-	if err != nil {
-		return err
-	}
-
-	wg.Add(1)
-	if r.serviceHost {
-		return instance.RunServiceHost(ctx, wg)
-	} else {
-		return instance.Run(ctx, wg)
-	}
 }
 
 func (r *serverRunCommand) prepareInstance() (i *server.Instance, err error) {
